+++ src/main/java/org/matsim/bene/analysis/eventsHandler/ParkingSlotVisualiserBus.java
@@ -1,4 +1,4 @@
-package org.matsim.bene.analysis;
+package org.matsim.bene.analysis.eventsHandler;
 
 import org.matsim.api.core.v01.Coord;
 import org.matsim.api.core.v01.Id;
